Rename hidinglist.ts to hiding_list.ts
