+++ src/neuroglancer/ui/hiding_list.ts
